Mostly finished accounts.py
The problem is that the node names change with addition of prices.
We need a way to traverse using only the item names.

One idea is to write a code which finds the price given an item.
Then we can deduce the new node names. However this will cause troubles
if there are two entries with the same name. We should discourage that.
May be in the add module, if a name exists, we add a number to it, like file
names.